requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
Brotli==1.1.0
cachetools==5.5.2
cashfree-pg==4.3.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
cssselect2==0.8.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
fonttools==4.58.0
frozenlist==1.5.0
google-auth==2.38.0
idna==3.10
multidict==6.4.3
pillow==11.1.0
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==1.10.22
pydyf==0.11.0
pyfcm==2.0.8
PyJWT==2.9.0
PyMySQL==1.1.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rsa==4.9
sentry-sdk==1.32.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.0.7
weasyprint==65.1
webencodings==0.5.1
yarl==1.19.0
zopfli==0.2.3.post1
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
mysqlclient>=2.1
python-decouple>=3.8
django-cors-headers>=4.0
cryptography>=41.0
django-environ==0.12.0
cryptograpy==0.0.0
pillow
pytz
requests


